refactor(backend): extract cardImages promotion pipeline into promotion/
chat/cardImages/ was 16 files and 7,448 lines flat. Move the four files of the durable
background promotion pipeline - driven by its own scheduled Lambda - into promotion/,
and leave the synchronous chat-tool request path (openaiAdapter, operation, contract,
types, metadata) flat where it belongs.
